_make_rescaled_ticks_and_units does not return None's anymore
Some commits ago _make_rescaled_ticks_and_units was updated to
return an exponential ticks formatter as a default (if the unit
is not in the list of those for which axes are rescaled and
the "engineering" prefixes are added). With this change, this
function can never return None's, hence the code needs to be
cleaned up to reflect that. This is what happens in this commit.