snap is a flexable class bound method
snap is now a class bound method.

it can be used as 
lockin.snap('x','y') -> tuple(x,y)
it checks for a correct number of parameters and that the parameters are valid.
It is not fussy about  'x' or  'X', and the confusing phase option can be 'p', 'P', 'Phase', 'phase', or 'θ'. to help with compatibility between similar drivers.