Dont convert non-grid dataframe to xarray - slow
Mikhail Astafev.

When building docs locally, the “dataset context manager”
example notebook takes too much time to execute to_xarray
in the scatter plot example.

is it because xarray is trying to make a grid out of 5000 ungridded points?

It's weird because the conversion is executed on a 10-rows
dataframe - a very small one. But the resulting xarray
has 5000 in its dimensions, as if it was made out of the
original huge dataframe.

I tried with "copy" - still the same problem. to_dict, to_json,
etc - work on a copy, no problem, but to_xarray goes nuts and
processes all the 5000 rows from the parent dataframe which it
should not know about since I made a “copy”.