Ensure instrument instance is closed on test cleanup
The thing is that the instrument instance will be created and will be
added to the registry of instrument instances thanks to the code of the
parent class of the 'S46'. However, in this test the __init__ of 'S46'
will break by intention which means that the instrument instance will
not be returned to the test (an exception will be raised) but that
instrument instance will still remain in the global registry of the
instrument instances which will cause a warning at the end of the test
execution. So, the way to remove that "almost created instance" of the
instrument is to use "close_all" class method (on the 'S46' instrument
class in order to be more specific with what needs to be closed).