Fix ami test
This test was broken in several ways.

1) The test compares the field rate limit in T/S against the current
rate limits in A/s. This means that ix.ramp_rate(target_ramp_rate)
did not actually raise the error tested for.
2) The test only asserted that any Exception was raised not checking
for a specific type of exception.
3) excinfo.value.args inside the context manager raises a AttributeError
which is then in turn caught by the pytest.raises context manager.
4) This in turn causes pytest to leave the tests in an inconsistent state
which results in other tests passing asserts that they should not.