Fix XArray FutureWarning
FutureWarning: The return type of `Dataset.dims` will be changed to
return a set of dimension names in future, in order to be more
consistent with `DataArray.dims`. To access a mapping from dimension
names to lengths, please use `Dataset.sizes`.

This fixes all places where we make use of .dims as a dict
There are a number of places where we use it to get the names/number of
the dimensions. Here we continue using .dims